+++ CHANGELOG.rst
@@ -45,6 +45,8 @@ Changed
 - Update ``errors`` dependency to 2.2.0.
 - Add default ``pool.map_call`` timeout 10 seconds.
 - Update ``frontend-core`` dependency to 7.8.0.
+- Argparse throws an error when it encouters ``instance_name`` missing in
+  instances.yml.
 
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Fixed